ollama_streamlit_demos-main/requirements.txt
ollama
streamlit==1.33.0
openai
chromadb
sentence_transformer
fitz